Add Khmer locale parity (#1757)
* feat(locale): add Khmer parity

* fix(locale): address Khmer review feedback

* fix(locale): stabilize Khmer decimal formatting

* fix(locale): make Khmer scale parsing general

* fix(locale): prefix joined-scale ordinal fallbacks

* fix(locale): parse glued Khmer scale counts

* fix(locale): group prior glued scale counts

* fix(locale): keep glued hundred groups current

* fix(locale): parse compact Khmer glued scale counts

* fix(locale): bound compact glued scale parsing

* fix(locale): fold compact glued scale groups

* fix(locale): cache compact glued scale tokens

* fix(locale): deduplicate compact glued scale tokens

+++ src/Humanizer/Localisation/NumberToWords/JoinedScaleNumberToWordsConverter.cs
@@ -134,7 +134,7 @@ class JoinedScaleNumberToWordsConverter(JoinedScaleNumberToWordsProfile profile)
         var block = profile.Ordinal!.Resolve(gender);
         if (number < 0)
         {
-            var magnitude = number == int.MinValue ? (long)int.MaxValue + 1 : Math.Abs((long)number);
+            var magnitude = number == int.MinValue ? (long)int.MaxValue + 1 : Math.Abs(number);
             if (magnitude <= int.MaxValue && block.ExactReplacements.TryGetValue((int)magnitude, out var negativeExact))
             {
                 return profile.MinusWord + profile.NegativeJoinWord + negativeExact;
@@ -309,7 +309,7 @@ internal sealed class JoinedScaleNumberToWordsProfile(
     }
 }
 
-/// <summary>Gender-specific ordinal affix and exact replacement data.</summary>
+/// <summary>Gender-specific ordinal affixes and exact replacement data.</summary>
 /// <param name="DefaultPrefix">The prefix prepended to the cardinal word form for productive ordinals.</param>
 /// <param name="DefaultSuffix">The suffix appended to the cardinal word form for productive ordinals.</param>
 /// <param name="ExactReplacements">Exact ordinal forms keyed by value.</param>
